# discobole-services/disco-catalog/catalog-product-catalog-administration/product-catalog-administration-service/src/main/java/com/orange/discobole/productcatalog/administration/controller/UnitController.java
// SPDX-FileCopyrightText: 2025 Orange SA
// SPDX-License-Identifier: MIT
//
// This software is distributed under the MIT License,
// the text of which is available at https://opensource.org/license/mit
// or see the "LICENSE.txt" file for more details.
//
// Authors: See CONTRIBUTORS.txt

package com.orange.discobole.productcatalog.administration.controller;


import com.orange.disco.admin.Unit;
import com.orange.discobole.productcatalog.administration.service.UnitService;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@Tag(name = "Unit", description = "Identifies the units of measure configured in the product catalog")
@Validated
public class UnitController {


    private UnitService unitService;

    @Autowired
    public UnitController(UnitService unitService) {
        this.unitService = unitService;
    }

    @PostMapping("/unit")
    public ResponseEntity<Unit> createUnit(@RequestBody(required = true) @Valid Unit unit) {
        return ResponseEntity.ok().body(this.unitService.createUnit(unit));
    }

    @GetMapping("/unit")
    public ResponseEntity <List< Unit >> getAllUnit() {
        return ResponseEntity.ok().body(unitService.getAllUnit());
    }

    @GetMapping("/unit/{id}")
    public ResponseEntity < Unit > getUnitById(@PathVariable String id) {
        return ResponseEntity.ok().body(unitService.getUnitById(id));
    }

    @PutMapping("/unit/{id}")
    public ResponseEntity < Unit > updateUnit(@PathVariable String id, @RequestBody Unit unit) {
        unit.setId(id);
        return ResponseEntity.ok().body(this.unitService.updateUnit(unit));
    }

    @DeleteMapping("/unit/{id}")
    public HttpStatus deleteUnit(@PathVariable String id) {
        this.unitService.deleteUnit(id);
        return HttpStatus.OK;
    }



}