# discobole-services/disco-catalog/catalog-product-catalog-administration/product-catalog-administration-spec/src/main/resources/catalog-administration-api-spec.yaml
...
openapi: "3.0.0"
info:
  description: "The productCatalogAdministration API provides for management which\
    \ includes CRUD operations for entities that are referred in product catalog like\
    \ catalog constants. These may be currencies, units of measures, frequencyTypes\
    \ etc. The API is not a TMF prescribed API and supports the following operations:\n\
    1.POST\n2.PATCH\n3.PUT\n4.DELETE"
  version: "5.7.0_inProgress"
  title: "ProductCatalogAdministration"
...
paths:
...
  /unit:
    post:
      tags:
        - "Unit"
      operationId: "unitCreate"
      summary: "create unit"
      requestBody:
        required: true
        content:
          application/json;charset=utf-8:
            schema:
              $ref: "#/components/schemas/unit"
      responses:
        "201":
          description: "Created"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/unit"
        "400":
          description: "Bad Request\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: "Method Not Allowed\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "Unprocessable entity\n\nBusiness error"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: "Service Unavailable\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - "Unit"
      operationId: "unitFind"
      summary: "find unit"
      responses:
        "200":
          description: "Ok"
          content:
            application/json;charset=utf-8:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/unit"
        "400":
          description: "Bad Request\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: "Method Not Allowed\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "Unprocessable entity\n\nBusiness error"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: "Service Unavailable\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
  /unit/{unitId}:
    get:
      tags:
        - "Unit"
      operationId: "unitGet"
      summary: "get unit"
      parameters:
        - name: "unitId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        "200":
          description: "Ok"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/unit"
        "400":
          description: "Bad Request\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: "Method Not Allowed\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "Unprocessable entity\n\nBusiness error"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: "Service Unavailable\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - "Unit"
      operationId: "unitUpdate"
      summary: "update unit"
      parameters:
        - name: "unitId"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json;charset=utf-8:
            schema:
              $ref: "#/components/schemas/unit"
      responses:
        "200":
          description: "Ok"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/unit"
        "400":
          description: "Bad Request\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: "Method Not Allowed\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "Unprocessable entity\n\nBusiness error"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: "Service Unavailable\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - "Unit"
      operationId: "unitDelete"
      summary: "delete unit"
      parameters:
        - name: "unitId"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        "204":
          description: "No Content"
        "400":
          description: "Bad Request\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Not Found\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: "Method Not Allowed\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "Unprocessable entity\n\nBusiness error"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Internal Server Error\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: "Service Unavailable\n\n"
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
...
servers:
  - url: "https://catalog-product-catalog-administration-integration-disco.apps.fr01.paas.tech.orange/productCatalogAdministration/v5/"
components:
  schemas:
    CFSRelationshipRestriction:
      description: "\n\n**Polymorphism**\n\n  Parent: productCatalogAdministration\n\
        \n  Discriminator: @type"
      allOf:
        - $ref: "#/components/schemas/productCatalogAdministration"
        - required:
            - "isRelationshipRestricted"
          type: "object"
          properties:
            isRelationshipRestricted:
              type: "boolean"
    Channel:
      description: "Object to define the channels supported for customer interactions.\n\
        \n**Polymorphism**\n\n  Parent: productCatalogAdministration\n\n  Discriminator:\
        \ @type"
      allOf:
        - $ref: "#/components/schemas/productCatalogAdministration"
        - required:
            - "name"
          type: "object"
          properties:
            name:
              description: "Name of the channel"
              type: "string"
    Error:
      description: "See [Orange recommendation](https://recommendations.innov.intraorange/designing-orange-api/#error-format)"
      required:
        - "code"
        - "message"
      type: "object"
      properties:
        code:
          description: "An integer coding the error type. This is given to caller\
            \ so he can translate them if required."
          type: "integer"
          format: "int32"
        message:
          description: "A short localized string that describes the error."
          type: "string"
        description:
          description: "(optional) A long localized error description if needed. It\
            \ can contain precise information about which parameter is missing, or\
            \ what are the identifier acceptable values."
          type: "string"
        infoURL:
          description: "(optional) A URL to online documentation that provides more\
            \ information about the error."
          type: "string"
        '@baseType':
          type: "string"
        schemaLocation:
          type: "string"
          format: "uri"
        '@referredType':
          type: "string"
    MarketSegment:
      description: "Describes the market segments defined in the product catalog\n\
        \n**Polymorphism**\n\n  Parent: productCatalogAdministration\n\n  Discriminator:\
        \ @type"
      allOf:
        - $ref: "#/components/schemas/productCatalogAdministration"
        - required:
            - "name"
          type: "object"
          properties:
            name:
              type: "string"
    ProductInventoryConfiguration:
      description: "Describe the configuration whether checks to product inventory\
        \ are to be instantiated.\n\n**Polymorphism**\n\n  Parent: productCatalogAdministration\n\
        \n  Discriminator: @type"
      allOf:
        - $ref: "#/components/schemas/productCatalogAdministration"
        - required:
            - "cpibCheck"
          type: "object"
          properties:
            cpibCheck:
              type: "boolean"
...
    currency:
      description: "This class is used to specify currencies for product catalog application.\
        \ The format follows the ISO4217 standard for currencies.\n\n**Polymorphism**\n\
        \n  Parent: productCatalogAdministration\n\n  Discriminator: @type"
      allOf:
        - $ref: "#/components/schemas/productCatalogAdministration"
        - required:
            - "label"
            - "code"
          type: "object"
          properties:
            label:
              description: "ISO4217 code for managing currencies"
              type: "string"
            code:
              description: "ISO4217 code for managing currencies"
              type: "string"
            isDefault:
              description: "Flag used by UI to indicate default selection"
              type: "boolean"
    frequencyTypes:
      description: "This class is used to identify frequency types associated with\
        \ time. These can be day(s), months(s) etc. which are to be used in the product\
        \ catalog while administering/creating product catalog entities.\n\n**Polymorphism**\n\
        \n  Parent: productCatalogAdministration\n\n  Discriminator: @type"
      allOf:
        - $ref: "#/components/schemas/productCatalogAdministration"
        - required:
            - "frequencyCode"
            - "frequencyLabel"
            - "numberOfDays"
          type: "object"
          properties:
            frequencyCode:
              type: "string"
            frequencyLabel:
              type: "string"
            numberOfDays:
              type: "integer"
              format: "int32"
    productCatalogAdministration:
      description: "\n\n**Polymorphism**\n\n  Discriminator: @type\n\n  One of:\n\
        \  - currency\n  - frequencyTypes\n  - unit\n  - ProductInventoryConfiguration\n\
        \  - Channel\n  - MarketSegment\n  - CFSRelationshipRestriction\n  - taxItem"
      discriminator:
        propertyName: "@type"
        mapping:
          currency: "#/components/schemas/currency"
          frequencyTypes: "#/components/schemas/frequencyTypes"
          unit: "#/components/schemas/unit"
          ProductInventoryConfiguration: "#/components/schemas/ProductInventoryConfiguration"
          Channel: "#/components/schemas/Channel"
          MarketSegment: "#/components/schemas/MarketSegment"
          CFSRelationshipRestriction: "#/components/schemas/CFSRelationshipRestriction"
          taxItem: "#/components/schemas/taxItem"
      required:
        - "lastUpdate"
      type: "object"
      properties:
        id:
          type: "string"
        '@type':
          type: "string"
        '@baseType':
          type: "string"
        schemaLocation:
          type: "string"
        href:
          type: "string"
        lastUpdate:
          type: "string"
          format: "date-time"
    taxItem:
      description: "\n\n**Polymorphism**\n\n  Parent: productCatalogAdministration\n\
        \n  Discriminator: @type"
      allOf:
        - $ref: "#/components/schemas/productCatalogAdministration"
        - required:
            - "description"
          type: "object"
          properties:
            taxAmount:
              type: "number"
              format: "float"
            taxRate:
              type: "integer"
              format: "int32"
            description:
              type: "string"
    unit:
      description: "The units of measurement describe the units for valued items.\
        \ Like Mb/s, GB/s etc.\n\n**Polymorphism**\n\n  Parent: productCatalogAdministration\n\
        \n  Discriminator: @type"
      allOf:
        - $ref: "#/components/schemas/productCatalogAdministration"
        - required:
            - "unitOfMeasure"
          type: "object"
          properties:
            unitOfMeasure:
              type: "string"